Update README.md
download the tax.sol file and deploy on remix ide. The smart contract saves the record of your name and asset information and finds the tax due of a particular person.in the add record functionality add your first name , last name, and asset, and then add record. You would be assigned a particular idno depending on how far you are in the record. you can check your tax due by entering your id no. The fact that the amount you in owe in tax is visible in the blockchain in public view creates accountability and hence could be used.